pyproject.toml
[project]
name = "wall-todo"
version = "0.1.0"
description = "E-paper todo board displaying Todoist tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "weasyprint>=60.0",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
]
raspi = [
    "waveshare-epaper>=1.0.0",
    "RPi.GPIO>=0.7.0",
    "spidev>=3.6",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
